+++ packaging/startos/actions/index.ts
@@ -1,4 +1,7 @@
 import { sdk } from '../sdk'
 import { rotatePassword } from './rotatePassword'
+import { selectIndexer } from './selectIndexer'
 
-export const actions = sdk.Actions.of().addAction(rotatePassword)
+export const actions = sdk.Actions.of()
+  .addAction(rotatePassword)
+  .addAction(selectIndexer)